Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@64196 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/objostrasn.cpp
@@ -161,7 +161,7 @@ void CObjectOStreamAsn::WriteUint8(Uint8 data)
     m_Output.PutUint8(data);
 }
 
-void CObjectOStreamAsn::WriteDouble2(double data, size_t digits)
+void CObjectOStreamAsn::WriteDouble2(double data, unsigned digits)
 {
 #if 1
     if (isnan(data)) {
